Rename Core.SharedResources to Core.Localization
Renamed the namespace `Core.SharedResources` to `Core.Localization` across multiple files for better representation of its purpose. Updated import statements accordingly. Made the `SharedResources` class public to allow external access.
